Added topics and posts page

+++ src/controllers/Topics.js
@@ -5,6 +5,20 @@ export async function getTopics() {
         .then(response => response.json())
 }
 
+export async function getTopicById(id) {
+    return fetch(`/topics/${id}`, {
+        method: 'GET',
+    })
+        .then(response => response.json())
+}
+
+export async function getPostsByTopicId(id) {
+    return fetch(`/topics/${id}/posts`, {
+        method: 'GET',
+    })
+        .then(response => response.json())
+}
+
 export async function createNewTopic() {
     const title = (document.getElementById("new-topic-title") || {}).value || ''
     const description = (document.getElementById("new-topic-description") || {}).value || ''
@@ -21,13 +35,13 @@ export async function createNewTopic() {
         }),
     })
 }
-export async function deleteTopicById() {
 
-    await fetch('/topics/', {
-        headers: {
-            'Content-Type': 'application/json'
-        },
-        method: 'DELETE',
-
-    })
+export async function deleteTopicById(id) {
+    const decision = confirm('Are you sure about deleting this topic?')
+    if (decision) {
+        await fetch(`/topics/${id}`, {
+            method: 'DELETE',
+        })
+    }
+    document.location.reload()
 }
